List

+++ app/src/main/java/com/tcp/thisable/Dao/Location.java
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName;
 
 public class Location {
     @SerializedName("type") public String type;
-    @SerializedName("coordinates") public float[] coordinates;
+    @SerializedName("coordinates") public double[] coordinates;
 }
